+++ src/git/GPI/src/Webhooks/Context/Components/Repository.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PCIT\GPI\Webhooks\Context\Components;
+
+/**
+ * @property string $default_branch
+ */
+class Repository
+{
+}