+++ src/git/GPI/src/Webhooks/Handler/Abstracts/PushAbstract.php
@@ -30,12 +30,13 @@ abstract class PushAbstract implements PushInterface
         $account = $context->account;
         $sender = $context->sender;
         $private = $context->private;
+        $default_branch = $context->repository->default_branch;
 
         // user table not include user info
         $subject = new Subject();
 
         $subject->register(
-            new UpdateUserInfo($account, (int) $installation_id, (int) $rid, $repo_full_name, $sender));
+            new UpdateUserInfo($account, (int) $installation_id, (int) $rid, $repo_full_name, $default_branch, $sender, $git_type));
 
         $config_array = $subject->register(new GetConfig((int) $rid, $commit_id, $git_type))->handle()->config_array;
 
@@ -67,11 +68,12 @@ abstract class PushAbstract implements PushInterface
         $account = $context->account;
         $sender = $context->sender;
         $private = $context->private;
+        $default_branch = $context->repository->default_branch;
 
         $subject = new Subject();
 
         $subject->register(
-            new UpdateUserInfo($account, (int) $installation_id, (int) $rid, $repo_full_name, $sender));
+            new UpdateUserInfo($account, (int) $installation_id, (int) $rid, $repo_full_name, $default_branch, $sender, $git_type));
 
         $config_array = $subject->register(new GetConfig((int) $rid, $commit_id, $git_type))->handle()->config_array;