+++ src/git/Gitee/src/Webhooks/Handler/Ping.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PCIT\Gitee\Webhooks\Handler;
+
+use PCIT\GPI\Webhooks\Handler\Abstracts\PingAbstract;
+
+class Ping extends PingAbstract
+{
+}